backend/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
pyyaml
python-multipart
pytest
pytest-asyncio
httpx
